item update in progress

+++ components/OrderCard.js
@@ -6,7 +6,9 @@ import { deleteOrder } from '../API/Promises';
 export default function OrderCard({ orderObj, onUpdate }) {
   const router = useRouter();
   const deleteThisOrder = () => {
-    deleteOrder(orderObj.id).then(() => onUpdate());
+    if (window.confirm('Delete Item?')) {
+      deleteOrder(orderObj.id).then(() => onUpdate());
+    }
   };
 
   const handleOrderClick = () => {
@@ -20,7 +22,7 @@ export default function OrderCard({ orderObj, onUpdate }) {
           <h2>{orderObj.name}</h2>
         </div>
         <div className="d-flex justify-content-between">
-          <button type="button">Edit</button>
+          <button type="button" onClick={() => router.push(`/create/${orderObj.id}`)}>Edit</button>
           <button type="button" onClick={handleOrderClick}>Details</button>
           <button type="button" onClick={() => router.push(`/item/${orderObj.id}`)}>Add Item</button>
           <button type="button" onClick={deleteThisOrder}>Delete</button>
